server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.1
openai>=1.30.0
httpx>=0.27.0
PyYAML>=6.0.1
python-multipart>=0.0.9
